requirement.txt
asgiref==3.7.2
Django==4.2.7
django-colorfield==0.10.1
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
mysqlclient==2.2.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
